+++ packages/lumx-react/src/components/thumbnail/index.ts
@@ -1,3 +1,2 @@
 export * from './Thumbnail';
 export * from './types';
-export { useFocusPointStyle } from '@lumx/react/components/thumbnail/useFocusPointStyle';